chore: Split NewAction into two classes
