+++ app/server/appsmith-interfaces/src/main/java/com/appsmith/external/models/ActionDTO.java
@@ -1,6 +1,7 @@
 package com.appsmith.external.models;
 
 import com.appsmith.external.models.ce.ActionCE_DTO;
+import com.querydsl.core.annotations.QueryEmbeddable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -10,4 +11,5 @@ import lombok.ToString;
 @Setter
 @NoArgsConstructor
 @ToString
+@QueryEmbeddable
 public class ActionDTO extends ActionCE_DTO {}
